refactor v2 collection op tests
